move code from http/admin to admin
